requirements.txt
Flask==1.1.0
maxminddb-geolite2==2018.703
